#679: Extract Redis store from server-cluster to new server-redis
