+++ leshan-integration-tests/pom.xml
@@ -33,7 +33,7 @@ Contributors:
         </dependency>
         <dependency>
             <groupId>org.eclipse.leshan</groupId>
-            <artifactId>leshan-server-cluster</artifactId>
+            <artifactId>leshan-server-redis</artifactId>
         </dependency>
         <dependency>
             <groupId>org.eclipse.leshan</groupId>
