+++ leshan-server-demo/pom.xml
@@ -38,7 +38,7 @@ Contributors:
         </dependency>
         <dependency>
             <groupId>org.eclipse.leshan</groupId>
-            <artifactId>leshan-server-cluster</artifactId>
+            <artifactId>leshan-server-redis</artifactId>
         </dependency>
         <dependency>
             <groupId>commons-cli</groupId>
