+++ leshan-server-redis/pom.xml
@@ -23,23 +23,15 @@ Contributors:
         <artifactId>leshan</artifactId>
         <version>1.0.0-SNAPSHOT</version>
     </parent>
-    <artifactId>leshan-server-cluster</artifactId>
+    <artifactId>leshan-server-redis</artifactId>
     <packaging>bundle</packaging>
-    <name>leshan - server cluster</name>
-    <description>Experimental clustering support for Leshan.</description>
+    <name>leshan - server redis</name>
+    <description>Redis implementation for different stores of Leshan Server</description>
 
     <dependencies>
         <dependency>
             <groupId>org.eclipse.leshan</groupId>
-            <artifactId>leshan-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.leshan</groupId>
             <artifactId>leshan-server-cf</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.leshan</groupId>
-            <artifactId>leshan-server-redis</artifactId>
         </dependency>
         <dependency>
             <groupId>redis.clients</groupId>
@@ -49,17 +41,6 @@ Contributors:
         <dependency>
             <groupId>com.eclipsesource.minimal-json</groupId>
             <artifactId>minimal-json</artifactId>
-        </dependency>
-         <dependency>
-            <groupId>commons-cli</groupId>
-            <artifactId>commons-cli</artifactId>
-        </dependency>
-
-        <!-- runtime dependencies -->
-        <dependency>
-            <groupId>ch.qos.logback</groupId>
-            <artifactId>logback-classic</artifactId>
-            <scope>runtime</scope>
         </dependency>
 
         <!-- test dependencies -->
@@ -83,28 +64,6 @@ Contributors:
                         <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
                     </instructions>
                 </configuration>
-            </plugin>
-            <plugin>
-                <artifactId>maven-assembly-plugin</artifactId>
-                <configuration>
-                    <archive>
-                        <manifest>
-                            <mainClass>org.eclipse.leshan.server.cluster.LeshanClusterServer</mainClass>
-                        </manifest>
-                    </archive>
-                    <descriptors>
-                        <descriptor>../jar-with-dependencies-with-logback.xml</descriptor>
-                    </descriptors>
-                </configuration>
-                <executions>
-                    <execution>
-                        <id>make-assembly</id> <!-- this is used for inheritance merges -->
-                        <phase>package</phase> <!-- bind to the packaging phase -->
-                        <goals>
-                            <goal>single</goal>
-                        </goals>
-                    </execution>
-                </executions>
             </plugin>
         </plugins>
     </build>
