+++ leshan-server-redis/src/main/java/org/eclipse/leshan/server/redis/RedisRegistrationStore.java
@@ -17,7 +17,7 @@
  *     Achim Kraus (Bosch Software Innovations GmbH) - update to modified 
  *                                                     ObservationStore API
  *******************************************************************************/
-package org.eclipse.leshan.server.cluster;
+package org.eclipse.leshan.server.redis;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
@@ -41,8 +41,8 @@ import org.eclipse.leshan.server.Startable;
 import org.eclipse.leshan.server.Stoppable;
 import org.eclipse.leshan.server.californium.CaliforniumRegistrationStore;
 import org.eclipse.leshan.server.californium.ObserveUtil;
-import org.eclipse.leshan.server.cluster.serialization.ObservationSerDes;
-import org.eclipse.leshan.server.cluster.serialization.RegistrationSerDes;
+import org.eclipse.leshan.server.redis.serialization.ObservationSerDes;
+import org.eclipse.leshan.server.redis.serialization.RegistrationSerDes;
 import org.eclipse.leshan.server.registration.Deregistration;
 import org.eclipse.leshan.server.registration.ExpirationListener;
 import org.eclipse.leshan.server.registration.Registration;
