+++ pom.xml
@@ -44,6 +44,7 @@ Contributors:
         <module>leshan-core-cf</module>
         <module>leshan-server-core</module>
         <module>leshan-server-cf</module>
+        <module>leshan-server-redis</module>
         <module>leshan-server-cluster</module>
         <module>leshan-client-core</module>
         <module>leshan-client-cf</module>
@@ -401,6 +402,11 @@ Contributors:
             <dependency>
                 <groupId>${project.groupId}</groupId>
                 <artifactId>leshan-client-cf</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>${project.groupId}</groupId>
+                <artifactId>leshan-server-redis</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
